services/cell_nats/requirements.txt
dbus-python
pynats2
